requirements.txt
# PDF Processing
pymupdf==1.23.8
pdfplumber==0.10.3
python-docx==1.1.0
PyPDF2==3.0.1

# Text Processing
regex==2023.12.25
python-dateutil==2.8.2

# Data Processing
pandas==2.1.4
numpy==1.26.2

# NLP & Embeddings
sentence-transformers==2.2.2

# Vector DB
chromadb==0.4.22
faiss-cpu==1.7.4

# LLM
google-generativeai==0.3.2

# Backend
fastapi==0.108.0
uvicorn==0.25.0
python-multipart==0.0.6

# Frontend
streamlit==1.29.0

# Utilities
python-dotenv==1.0.0
pydantic==2.5.3
tqdm==4.66.1
loguru==0.7.3

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# OCR support (for scanned PDFs)
pytesseract==0.3.10
pillow==10.2.0